emulator: adding cleaner emulator hierarchy
